Move traits and classes to own files. Document.
Moved traits and classes to own files grouped by incumbencies.

Started documenting main traits, classes and methods.
