+++ src/main/scala/org/retistruen/Emitter.scala
@@ -1,0 +1,74 @@
+package org.retistruen
+
+import org.joda.time.Instant
+
+/** Emits values as instances of $Datum to the set of registered $Receiver
+ *  @tparam T the type of the emitted values
+ *  @define Receiver [[org.retistruen.Receiver]]
+ *  @define Emitter [[org.retistruen.Emitter]]
+ *  @define Datum [[org.retistruen.Datum]]
+ *  @define Emits Emits the value wrapped in the given $Datum to all registered $Receiver
+ *  @define Registers Registers the $Receiver to be sent values emitted by this Emitter.
+ *  @define EmitsValue Emits the value provided wrapped in a new $Datum
+ */
+trait Emitter[T] extends Named {
+
+  private var receivers: Set[Receiver[T]] = Set.empty
+
+  /** Returns the set of $Receiver registered with this Emitter to receive emitted $Datum */
+  def registered: Set[Receiver[T]] = receivers
+
+  /** $Registers */
+  def register(rec: Receiver[T]): Unit =
+    receivers = receivers + rec
+
+  /** $Registers */
+  def >>(other: Receiver[T]): Unit = register(other)
+
+  /** $Registers */
+  def >>>(others: Receiver[T]*): Unit = others.foreach(register(_))
+
+  /** $Registers Also returns the {{$Receiver with Emitter}} as Emitter to
+   *  support a fluent registration idiom like:
+   *  {{{
+   *  emitter1 >> emitterreceiver1 >> emitterreceiver2 >> receiver
+   *  }}}
+   */
+  def >>(other: Receiver[T] with Emitter[T]): Emitter[T] = {
+    register(other)
+    other
+  }
+
+  /** $Emits */
+  def emit(datum: Datum[T]): Unit =
+    receivers.foreach(_.receive(this, datum))
+
+  /** $EmitsValue */
+  def emit(some: T): Unit =
+    emit(Datum(some, new Instant))
+
+  /** $Emits */
+  def <<(datum: Datum[T]) = emit(datum)
+
+  /** $EmitsValue */
+  def <<(some: T) = emit(some)
+
+  /** $EmitsValue */
+  def <<<(some: T*) = some foreach emit
+
+}
+
+/** $Emitter that keeps the last emitted $Datum cached for further access */
+trait CachingEmitter[T] extends Emitter[T] {
+
+  private var cached: Option[Datum[T]] = None
+
+  /** Returns the last $Datum emitted by this $Emitter */
+  def last: Option[Datum[T]] = cached
+
+  override def emit(datum: Datum[T]) = {
+    cached = Some(datum)
+    super.emit(datum)
+  }
+
+}
