+++ src/main/scala/org/retistruen/Model.scala
@@ -1,0 +1,5 @@
+package org.retistruen
+
+import org.retistruen.dsl.ModelBuilding
+
+class Model(val name: String) extends Named with ModelBuilding
