+++ src/main/scala/org/retistruen/Named.scala
@@ -1,0 +1,11 @@
+package org.retistruen
+
+/** Common trait for every object which has a name */
+trait Named {
+
+  val name: String
+
+  override def toString =
+    "%s[%s]" format (getClass.getSimpleName, name)
+
+}
