+++ src/main/scala/org/retistruen/dsl/package.scala
@@ -1,0 +1,4 @@
+package org.retistruen
+
+/** DSL for building [[org.retistruen.Model]] idiomatically */
+package object dsl
