+++ src/main/scala/org/retistruen/instrument/package.scala
@@ -1,0 +1,4 @@
+package org.retistruen
+
+/** Contains usable instruments ready to be connected in models */
+package object instrument
