+++ src/main/scala/org/retistruen/view/package.scala
@@ -1,0 +1,6 @@
+package org.retistruen
+
+/** Contains a simple {org.retistruen.Model} viewer
+ *  @see {org.reristruen.view.ModelViewer} documentatino
+ */
+package object view
