--- src/test/scala/org/retistruen/AsynchronousFun.scala
@@ -1,20 +1,0 @@
-/*
- * Author: Pablo Lalloni <plalloni@gmail.com>
- * Created: 24/05/2011 18:06:55
- */
-package org.retistruen
-
-import org.scalatest.FunSuite
-import scala.actors.Actor
-
-/**
- * @author Pablo Lalloni <plalloni@gmail.com>
- * @since 24/05/2011 18:06:55
- */
-class AsynchronousSuite extends FunSuite {
-
-  test("asynchrounous emit") {
-    pending
-  }
-
-}
